+++ ShareClient.php
@@ -72,12 +72,15 @@ final readonly class ShareClient
             throw new UnableToGenerateSasException;
         }
 
+        $builder = clone $shareSasBuilder;
+
         if (StorageUriParserHelper::isDevelopmentUri($this->uri)) {
-            $shareSasBuilder->setProtocol(SasProtocol::HTTPS_AND_HTTP);
+            $builder->setProtocol(SasProtocol::HTTPS_AND_HTTP);
         }
 
-        $sas = $shareSasBuilder
+        $sas = $builder
             ->setShareName($this->shareName)
+            ->setFilePath(null)
             ->build($this->credential);
 
         return $this->uri->withQuery(Query::build([